Update to iOS 7+
Updated project to build with iOS 7+ devices.

Eliminated any warnings/compiler errors created; many of them related
to image resources sizes for loading screens and app icons.

Added to .psd Photoshop file for some of the graphics.

+++ Inner Life/Inner Life/ILDashboardViewController.m
@@ -174,7 +174,7 @@
                          NSString *s = [[NSString alloc] initWithData:body encoding:NSUTF8StringEncoding];
                          NSLog(@"Web API Response:\n%@", s);
                          
-                         int http_status_code = [((NSHTTPURLResponse *)response) statusCode];
+                         NSInteger http_status_code = [((NSHTTPURLResponse *)response) statusCode];
                          
                          // HTTP (201 Created).
                          if (http_status_code == 201)
